chore(common/compiled): move Wss class from services/ to ws/ and remove stale auth files
